server/requirements.txt
fastapi
uvicorn[standard]
orjson
websockets
tortoise-orm[asyncpg]
python-rapidjson
ciso8601
ulid-py
